+++ api/repository/repository_users.go
@@ -4,10 +4,10 @@ import "apartment/api/models"
 
 type UserRepository interface {
 	FindAll() ([]models.User, error)
-	FindByID(uint32) (models.User, error)
+	FindByID(uint64) (models.User, error)
 	Create(models.User) (models.User, error)
-	Update(uint32, models.User) (int64, error)
-	Delete(uint32) (int64, error)
+	Update(uint64, models.User) (int64, error)
+	Delete(uint64) (int64, error)
 
 	CheckUniqueUser(models.User) (bool, error)
 }
